add new ALLOW_TOOLBAR_CALLBACK setting
Commit 68c4410f01cb31d199fc4aed86db9bdf965ccbad changed the semantics of
SHOW_TOOLBAR_CALLBACK to mean whether or not access to the toolbar
should be allowed; not just if the toolbar should be displayed or not.
This commit restores the original meaning of SHOW_TOOLBAR_CALLBACK and
adds a new setting, ALLOW_TOOLBAR_CALLBACK, which encapsulates the
functionality implemented by 68c4410f01cb31d199fc4aed86db9bdf965ccbad.
In conjunction with this, the require_show_toolbar() decorator was
replaced with a new require_allow_toolbar() decorator.